+++ spec/slack/web/api/mixins/groups_spec.rb
@@ -2,7 +2,7 @@
 require 'spec_helper'
 
 RSpec.describe Slack::Web::Api::Mixins::Groups do
-  subject do
+  subject(:groups) do
     klass.new
   end
 
@@ -13,7 +13,7 @@ RSpec.describe Slack::Web::Api::Mixins::Groups do
   end
 
   before do
-    allow(subject).to receive(:groups_list).and_return(
+    allow(groups).to receive(:groups_list).and_return(
       Slack::Messages::Message.new(
         'groups' => [{
           'id' => 'CDEADBEEF',
@@ -25,17 +25,17 @@ RSpec.describe Slack::Web::Api::Mixins::Groups do
 
   context '#groups_id' do
     it 'leaves groups specified by ID alone' do
-      expect(subject.groups_id(channel: 'C123456')).to(
+      expect(groups.groups_id(channel: 'C123456')).to(
         eq('ok' => true, 'group' => { 'id' => 'C123456' })
       )
     end
     it 'translates a channel that starts with a #' do
-      expect(subject.groups_id(channel: '#general')).to(
+      expect(groups.groups_id(channel: '#general')).to(
         eq('ok' => true, 'group' => { 'id' => 'CDEADBEEF' })
       )
     end
     it 'fails with an exception' do
-      expect { subject.groups_id(channel: '#invalid') }.to(
+      expect { groups.groups_id(channel: '#invalid') }.to(
         raise_error(Slack::Web::Api::Errors::SlackError, 'channel_not_found')
       )
     end